renames files and merge s3x_test.go with gateway prefixed test
