+++ gateway/s3x/errors.go
