+++ gateway/s3x/object.go
