+++ gateway/s3x/object_test.go
