+++ gateway/s3x/s3x_utils.go
